requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
chromadb==0.4.18
langchain==0.0.335
langchain-openai==0.0.2
arxiv==2.1.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.1
pypdf==3.17.1
sentence-transformers==2.2.2
